Extract `rendering`
