+++ src/rendering/constants.ts
@@ -1,0 +1,12 @@
+export enum PRIORITY {
+  DOM_READ = 'DOM_READ',
+  COMPUTATION = 'COMPUTATION',
+  DOM_WRITE = 'DOM_WRITE',
+  FUTURE = 'FUTURE',
+}
+
+export const ORDER = [
+  PRIORITY.DOM_READ,
+  PRIORITY.COMPUTATION,
+  PRIORITY.DOM_WRITE,
+]
